Add documentation and general cleanup.
